Bugfix: support current stable i3 con "type"
In the current stable version of i3, the con `type` property is an
integer that indicates type from an enum in the data header.

When initializing a con, if the json object `type` member is found to be
an int, it will be converted to the new string property internally.