fix memory leak of i3ipcCon
i3ipcCon is a tree and each node keeps a reference to its parent and
children. Because of the circular reference, the nodes in the tree were
never freed.

Therefore, make the child only hold a weak reference to its parent
instead of calling g_object_ref. Assure that the parent field is unset
whenever the parent of a separately referenced child is freed.